fix: Fix changes after merge
