+++ src/components/AuditLogsTable/AuditLogsTableActions/AuditLogsTableActions.tsx
@@ -1,17 +1,6 @@
-import {
-  Button,
-  Icon,
-  Tooltip,
-  Pagination,
-  Select,
-} from "@canonical/react-components";
-import { useCallback, type OptionHTMLAttributes } from "react";
-import { useSelector } from "react-redux";
+import { Button, Icon, Tooltip } from "@canonical/react-components";
 
 import { useQueryParams } from "hooks/useQueryParams";
-import { actions as jujuActions } from "store/juju";
-import { getAuditEvents, getAuditEventsLimit } from "store/juju/selectors";
-import { useAppDispatch } from "store/store";
 
 import AuditLogsTablePagination from "../AuditLogsTablePagination";
 import { DEFAULT_PAGE } from "../consts";
@@ -23,26 +12,14 @@ export enum Label {
 }
 
 const AuditLogsTableActions = () => {
-  const dispatch = useAppDispatch();
-  const auditLogs = useSelector(getAuditEvents);
-  const [queryParams, setQueryParams] = useQueryParams<{
+  const [, setQueryParams] = useQueryParams<{
     panel: string | null;
     page: string | null;
   }>({
     panel: null,
     page: DEFAULT_PAGE,
   });
-  const limit = useSelector(getAuditEventsLimit);
-  const page = Number(queryParams.page);
   const fetchAuditEvents = useFetchAuditEvents();
-
-  const handleChangeSelect = useCallback(
-    (e: React.ChangeEvent<HTMLSelectElement>) => {
-      setQueryParams({ page: null }, { replace: true });
-      dispatch(jujuActions.updateAuditEventsLimit(Number(e.target.value)));
-    },
-    [dispatch, setQueryParams]
-  );
 
   return (
     <>
